Clean up create_basic_fake_processor definitions and usage
We had this function in three different places and one of those instances was
slightly different than the other two. Seems like it helps to have a centralized
"fake processor" configuration that is used everywhere.

This centralizes that code and updates the other files to use it.
