+++ socorro/unittest/processor/test_skunk_classifiers.py
@@ -6,7 +6,7 @@ import copy
 
 from nose.tools import eq_, ok_
 
-from socorro.lib.util import DotDict, SilentFakeLogger
+from socorro.lib.util import DotDict
 from socorro.processor.skunk_classifiers import (
     SkunkClassificationRule,
     DontConsiderTheseFilter,
@@ -16,30 +16,14 @@ from socorro.processor.skunk_classifiers import (
     Bug811804,
     Bug812318,
 )
-from socorro.processor.signature_utilities import CSignatureTool
+from socorro.unittest.processor import (
+    create_basic_fake_processor,
+    c_signature_tool,
+)
 from socorro.unittest.processor.test_breakpad_pipe_to_json import (
     cannonical_json_dump
 )
 from socorro.unittest.testbase import TestCase
-
-csig_config = DotDict()
-csig_config.irrelevant_signature_re = ''
-csig_config.prefix_signature_re = ''
-csig_config.signatures_with_line_numbers_re = ''
-csig_config.signature_sentinels = []
-csig_config.collapse_arguments = True
-c_signature_tool = CSignatureTool(csig_config)
-
-
-def create_basic_fake_processor():
-    fake_processor = DotDict()
-    fake_processor.c_signature_tool = c_signature_tool
-    fake_processor.config = DotDict()
-    # need help figuring out failures? switch to FakeLogger and read stdout
-    fake_processor.config.logger = SilentFakeLogger()
-    fake_processor.processor_notes = []
-    #fake_processor.config.logger = FakeLogger()
-    return fake_processor
 
 
 class TestSkunkClassificationRule(TestCase):
