+++ socorro/unittest/processor/test_support_classifiers.py
@@ -8,35 +8,18 @@ from nose.tools import eq_, ok_
 
 from sys import maxint
 
-from socorro.lib.util import DotDict, SilentFakeLogger
+from socorro.lib.util import DotDict
 from socorro.processor.support_classifiers import (
     SupportClassificationRule,
     BitguardClassifier,
     OutOfDateClassifier,
 )
 
-from socorro.processor.signature_utilities import CSignatureTool
+from socorro.unittest.processor import create_basic_fake_processor
 from socorro.unittest.processor.test_breakpad_pipe_to_json import (
     cannonical_json_dump,
 )
 from socorro.unittest.testbase import TestCase
-
-csig_config = DotDict()
-csig_config.irrelevant_signature_re = ''
-csig_config.prefix_signature_re = ''
-csig_config.signatures_with_line_numbers_re = ''
-csig_config.signature_sentinels = []
-c_signature_tool = CSignatureTool(csig_config)
-
-
-def create_basic_fake_processor():
-    fake_processor = DotDict()
-    fake_processor.c_signature_tool = c_signature_tool
-    fake_processor.config = DotDict()
-    # need help figuring out failures? switch to FakeLogger and read stdout
-    fake_processor.config.logger = SilentFakeLogger()
-    #fake_processor.config.logger = FakeLogger()
-    return fake_processor
 
 
 class TestSupportClassificationRule(TestCase):
